feat: Auth hoc
