+++ client/src/App.tsx
@@ -1,67 +1,34 @@
-import React, { useState, useEffect } from 'react'
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  useHistory,
-} from 'react-router-dom'
+import React, { useEffect } from 'react'
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import { createGlobalStyle } from 'styled-components'
-import checkUserToken from '@api/user'
 import { WorkspaceJoinPage, ChannelPage, LoginPage, WorkspacePage } from '@page'
+import { ToastContainer } from 'react-toastify'
+import 'react-toastify/dist/ReactToastify.css'
+import Auth from '@hook/Auth'
 
 const App = () => {
-  const token = localStorage.getItem('token')
-  const [isAuth, setIsAuth] = useState<boolean>(false)
-  const history = useHistory()
-
-  const [name, accessToken] = window.location.search.split('=')
-
-  if (accessToken && name === '?access_token') {
-    localStorage.setItem('token', accessToken)
-    window.location.href = '/'
-  }
-
-  const checkToken = async (): Promise<boolean> => {
-    try {
-      const { success } = await checkUserToken()
-      if (success) {
-        setIsAuth(true)
-        return true
-      }
-      return false
-    } catch (error) {
-      console.log(error)
-      return false
-    }
-  }
-
-  const checkUser = async () => {
-    if (!token) {
-      history.push('/login')
-    }
-    if (token) {
-      try {
-        await checkToken()
-        history.push('/')
-      } catch (err) {
-        history.push('/login')
-      }
-    }
-  }
-
   useEffect(() => {
-    checkUser()
-  }, [])
+    const [name, accessToken] = window.location.search.split('=')
 
+    if (accessToken && name === '?access_token') {
+      localStorage.setItem('token', accessToken)
+      window.location.href = '/'
+    }
+  })
   return (
     <>
+      <ToastContainer />
       <GlobalStyle />
       <Router>
         <Switch>
-          <Route exact path="/" component={WorkspacePage} />
-          <Route exact path="/login" component={LoginPage} />
-          <Route exact path="/workspace_join" component={WorkspaceJoinPage} />
-          <Route exact path="/channel" component={ChannelPage} />
+          <Route exact path="/" component={Auth(WorkspacePage, false)} />
+          <Route exact path="/login" component={Auth(LoginPage, true)} />
+          <Route
+            exact
+            path="/workspace_join"
+            component={Auth(WorkspaceJoinPage, false)}
+          />
+          <Route exact path="/channel" component={Auth(ChannelPage, false)} />
         </Switch>
       </Router>
     </>
