+++ client/src/hook/Auth.tsx
@@ -1,0 +1,35 @@
+import React, { useEffect, useState } from 'react'
+import { toast } from 'react-toastify'
+import checkUserToken from '@api/user'
+import { useHistory } from 'react-router-dom'
+
+const Auth = (Component: any, option: boolean) => () => {
+  const history = useHistory()
+  const token = localStorage.getItem('token')
+  useEffect(() => {
+    const check = async () => {
+      try {
+        if (token) {
+          const { success } = await checkUserToken()
+          if (success && option) {
+            history.push('/login')
+          }
+          if (success && !option) {
+            history.push('/')
+          }
+        } else {
+          history.push('/login')
+        }
+      } catch (err) {
+        toast.error('잘못된 접근입니다.', {
+          onClose: () => history.push('/login'),
+        })
+      }
+    }
+    check()
+  }, [])
+
+  return <Component />
+}
+
+export default Auth
